Move TwitchClient into separate directory, extract EventArgs
