+++ JumpRoyale/src/Twitch/TwitchClient.cs
@@ -10,19 +10,6 @@ using TwitchLib.PubSub.Events;
 
 namespace TwitchChat
 {
-    public class MessageEventArgs : EventArgs
-    {
-        public string Message { get; set; } = string.Empty;
-
-        public string SenderName { get; set; } = string.Empty;
-
-        public string SenderId { get; set; } = string.Empty;
-
-        public string HexColor { get; set; } = string.Empty;
-
-        public bool IsPrivileged { get; set; }
-    }
-
     public class TwitchChatClient
     {
         /// <summary>
